requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.1
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
pytest-mock==3.12.0
